# docs/swagger.yaml
basePath: /api
definitions:
  ds.Lamp:
    properties:
      id:
        type: integer
      image_url:
        type: string
      luminous_flux_lm:
        type: number
      power_w:
        type: number
      scattering_angle_deg:
        type: number
      title:
        type: string
    type: object
  ds.LightRequest:
    properties:
      closed_at:
        type: string
      created_at:
        type: string
      formed_at:
        type: string
      light_request_to_lamp:
        items:
          $ref: '#/definitions/ds.LightRequestToLamp'
        type: array
      max_total_power_w:
        type: number
      status:
        description: 1 - draft, 2 - deleted, 3 - pending, 4 - resolved, 5 - rejected
        type: integer
    type: object
  ds.LightRequestToLamp:
    properties:
      area_m2:
        type: number
      lamp:
        $ref: '#/definitions/ds.Lamp'
      number:
        type: integer
    type: object
...
  handler.CartInfoResponse:
    properties:
      item_count:
        type: integer
      request_id:
        type: integer
    type: object
...
  handler.UpdateRequestRequest:
    properties:
      max_total_power_w:
        type: number
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourusername/dia-backend
  description: REST API для веб-приложения расчета освещения комнат
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Let There Be Light API
  version: "1.0"
paths:
...
  /light-requests:
    get:
      consumes:
      - application/json
      description: Get all light requests with optional filters (moderator only)
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: integer
      - description: Filter from date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter to date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.LightRequest'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all light requests
      tags:
      - light-requests
  /light-requests/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a draft light request
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete light request
      tags:
      - light-requests
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific light request
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.LightRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get light request by ID
      tags:
      - light-requests
    put:
      consumes:
      - application/json
      description: Update max total power for a draft light request
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update light request
      tags:
      - light-requests
  /light-requests/{id}/form:
    put:
      consumes:
      - application/json
      description: Submit draft request for moderation
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Form light request
      tags:
      - light-requests
  /light-requests/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a light request (moderator only)
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject light request
      tags:
      - light-requests
  /light-requests/{id}/resolve:
    put:
      consumes:
      - application/json
      description: Approve and resolve a light request with calculations (moderator
        only)
      parameters:
      - description: Light Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resolve light request
      tags:
      - light-requests
  /light-requests/cart:
    get:
      consumes:
      - application/json
      description: Get draft request ID and item count for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CartInfoResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cart information
      tags:
      - light-requests
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"